+++ CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Add years expansion to preprocessor.
 - Add excerpt cleaner.
+### Fixed
+- Fix double playback issue in debugger
 ## [0.2.1] 2026-06-02
 ### Fixed
 - Show proper version of Narrative in the main view.